Made excluded plugins configurable
Using configable mod role in debug command
Also moved the AnnouncerModule
