+++ src/main/java/buttondevteam/discordplugin/mcchat/MCChatUtils.java
@@ -8,17 +8,25 @@ import buttondevteam.lib.TBMCSystemChatEvent;
 import io.netty.util.collection.LongObjectHashMap;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.var;
+import lombok.val;
 import org.bukkit.Bukkit;
 import org.bukkit.command.CommandSender;
+import org.bukkit.event.Event;
+import org.bukkit.event.HandlerList;
+import org.bukkit.plugin.AuthorNagException;
+import org.bukkit.plugin.Plugin;
+import org.bukkit.plugin.RegisteredListener;
 import sx.blah.discord.handle.obj.IChannel;
 import sx.blah.discord.handle.obj.IMessage;
 import sx.blah.discord.handle.obj.IUser;
 
 import javax.annotation.Nullable;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Optional;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
+import java.util.logging.Level;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -211,6 +219,69 @@ public class MCChatUtils {
 			}
 		}
 		//If it gets here, it's sending a message to a non-chat channel
+	}
+
+	public static void callEventExcludingSome(Event event) {
+		if (notEnabled()) return;
+		callEventExcluding(event, false, module.excludedPlugins().get());
+	}
+
+	/**
+	 * Calls an event with the given details.
+	 * <p>
+	 * This method only synchronizes when the event is not asynchronous.
+	 *
+	 * @param event   Event details
+	 * @param only    Flips the operation and <b>includes</b> the listed plugins
+	 * @param plugins The plugins to exclude. Not case sensitive.
+	 */
+	@SuppressWarnings("WeakerAccess")
+	public static void callEventExcluding(Event event, boolean only, String... plugins) { // Copied from Spigot-API and modified a bit
+		if (event.isAsynchronous()) {
+			if (Thread.holdsLock(Bukkit.getPluginManager())) {
+				throw new IllegalStateException(
+					event.getEventName() + " cannot be triggered asynchronously from inside synchronized code.");
+			}
+			if (Bukkit.getServer().isPrimaryThread()) {
+				throw new IllegalStateException(
+					event.getEventName() + " cannot be triggered asynchronously from primary server thread.");
+			}
+			fireEventExcluding(event, only, plugins);
+		} else {
+			synchronized (Bukkit.getPluginManager()) {
+				fireEventExcluding(event, only, plugins);
+			}
+		}
+	}
+
+	private static void fireEventExcluding(Event event, boolean only, String... plugins) {
+		HandlerList handlers = event.getHandlers(); // Code taken from SimplePluginManager in Spigot-API
+		RegisteredListener[] listeners = handlers.getRegisteredListeners();
+		val server = Bukkit.getServer();
+
+		for (RegisteredListener registration : listeners) {
+			if (!registration.getPlugin().isEnabled()
+				|| Arrays.stream(plugins).anyMatch(p -> only ^ p.equalsIgnoreCase(registration.getPlugin().getName())))
+				continue; // Modified to exclude plugins
+
+			try {
+				registration.callEvent(event);
+			} catch (AuthorNagException ex) {
+				Plugin plugin = registration.getPlugin();
+
+				if (plugin.isNaggable()) {
+					plugin.setNaggable(false);
+
+					server.getLogger().log(Level.SEVERE,
+						String.format("Nag author(s): '%s' of '%s' about the following: %s",
+							plugin.getDescription().getAuthors(), plugin.getDescription().getFullName(),
+							ex.getMessage()));
+				}
+			} catch (Throwable ex) {
+				server.getLogger().log(Level.SEVERE, "Could not pass event " + event.getEventName() + " to "
+					+ registration.getPlugin().getDescription().getFullName(), ex);
+			}
+		}
 	}
 
 	@RequiredArgsConstructor
